# MessagePack Formatter: Cache Span tag keys in UTF8