#### Old:

// * Summary *

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Xeon CPU E5-2673 v3 2.40GHz, 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=5.0.201
  [Host]     : .NET Core 3.1.7 (CoreCLR 4.700.20.36602, CoreFX 4.700.20.37001), X64 RyuJIT
  Job-WEIHTL : .NET Framework 4.8 (4.8.4341.0), X64 RyuJIT
  Job-VTKECX : .NET Core 3.1.7 (CoreCLR 4.700.20.36602, CoreFX 4.700.20.37001), X64 RyuJIT

IterationTime=1.5000 s  

|                      Method |        Job |       Runtime |     Toolchain |     Mean |    Error |   StdDev | Ratio | RatioSD | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------------- |----------- |-------------- |-------------- |---------:|---------:|---------:|------:|--------:|------:|------:|------:|----------:|
|         WriteAndFlushTraces | Job-WEIHTL |    .NET 4.7.2 |        net472 | 748.0 us | 10.84 us |  9.61 us |  1.00 |    0.00 |     - |     - |     - |   3.03 KB |
|         WriteAndFlushTraces | Job-VTKECX | .NET Core 3.1 | netcoreapp3.1 | 604.5 us | 11.94 us | 12.78 us |  0.81 |    0.02 |     - |     - |     - |   2.45 KB |
|                             |            |               |               |          |          |          |       |         |       |       |       |           |
| WriteAndFlushEnrichedTraces | Job-WEIHTL |    .NET 4.7.2 |        net472 | 937.3 us | 13.01 us | 11.54 us |  1.00 |    0.00 |     - |     - |     - |   3.04 KB |
| WriteAndFlushEnrichedTraces | Job-VTKECX | .NET Core 3.1 | netcoreapp3.1 | 743.7 us | 14.43 us | 16.04 us |  0.79 |    0.02 |     - |     - |     - |   2.45 KB |


#### New:

// * Summary *

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Xeon CPU E5-2673 v3 2.40GHz, 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=5.0.201
  [Host]     : .NET Core 3.1.7 (CoreCLR 4.700.20.36602, CoreFX 4.700.20.37001), X64 RyuJIT
  Job-GAZNKL : .NET Framework 4.8 (4.8.4341.0), X64 RyuJIT
  Job-NQGFHJ : .NET Core 3.1.7 (CoreCLR 4.700.20.36602, CoreFX 4.700.20.37001), X64 RyuJIT

IterationTime=1.5000 s  

|                      Method |        Job |       Runtime |     Toolchain |     Mean |    Error |   StdDev | Ratio | RatioSD | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------------- |----------- |-------------- |-------------- |---------:|---------:|---------:|------:|--------:|------:|------:|------:|----------:|
|         WriteAndFlushTraces | Job-GAZNKL |    .NET 4.7.2 |        net472 | 604.9 us |  6.84 us |  6.06 us |  1.00 |    0.00 |     - |     - |     - |   3.03 KB |
|         WriteAndFlushTraces | Job-NQGFHJ | .NET Core 3.1 | netcoreapp3.1 | 467.7 us |  6.88 us |  5.74 us |  0.77 |    0.01 |     - |     - |     - |   2.45 KB |
|                             |            |               |               |          |          |          |       |         |       |       |       |           |
| WriteAndFlushEnrichedTraces | Job-GAZNKL |    .NET 4.7.2 |        net472 | 819.3 us | 15.33 us | 14.34 us |  1.00 |    0.00 |     - |     - |     - |   3.03 KB |
| WriteAndFlushEnrichedTraces | Job-NQGFHJ | .NET Core 3.1 | netcoreapp3.1 | 624.5 us |  7.92 us |  6.18 us |  0.76 |    0.02 |     - |     - |     - |   2.45 KB |


#### Diff:

1) WriteAndFlushTraces/net472: -143.1 us
2) WriteAndFlushTraces/netcoreapp3.1: -136,8 us
3) WriteAndFlushEnrichedTraces/net472: -118 us
4) WriteAndFlushEnrichedTraces/netcoreapp3.1: -119,2 us

![image](https://user-images.githubusercontent.com/69803/119346624-bdc93e00-bc9a-11eb-8d73-d6d449597d6d.png)


@DataDog/apm-dotnet