@tonyredondo 2021-05-24
> Nice 👍 I'm kind of surprised there's no change in the allocations though, I guess `MessagePackBinary.WriteString` is caching the `byte[]` behind the scene?

No, is not caching, is writing directly to the actual buffer... https://github.com/DataDog/dd-trace-dotnet/blob/master/src/Datadog.Trace/Vendors/MessagePack/MessagePackBinary.cs#L1715

Also, I do think there is a change on allocations, can be seen in `WriteAndFlushEnrichedTraces/net472` from `3.04 KB` to `3.03 KB`. But is a little change compared to all allocations.